Updated Base class for Language Model Experiment and Refactored IMDB Experiment
